+++ src/clj/rems/db/csv.clj
@@ -8,18 +8,28 @@
             [rems.db.user-settings :as user-settings]
             [rems.text :as text]))
 
+(def ^:private crlf "\r\n")
+
 (defn print-to-csv [& {:keys [column-names rows
-                              quote-strings? separator]
+                              quote-strings? separator
+                              strip-line-returns?]
                        :or {separator (:csv-separator env)}}]
   (let [escape-quotes #(str/replace % "\"" "\\\"")
+        strip-line-returns #(str/replace % #"[\r\n]" " ")
+        maybe-strip #(if (and (string? %)
+                             strip-line-returns?)
+                       (strip-line-returns %)
+                       %)
         maybe-quote #(if (and (string? %)
                               quote-strings?)
                        (str "\"" (escape-quotes %) "\"")
                        %)]
     (with-out-str
-      (println (str/join separator (mapv maybe-quote column-names)))
+      (print (str/join separator (mapv (comp maybe-quote maybe-strip) column-names)))
+      (print crlf)
       (doseq [row rows]
-        (println (str/join separator (mapv maybe-quote row)))))))
+        (print (str/join separator (mapv (comp maybe-quote maybe-strip) row)))
+        (print crlf)))))
 
 ;; Export applications
 
@@ -74,25 +84,23 @@
 
 (defn- form-field-names [form-id application]
   (let [form (find-first #(= form-id (:form/id %)) (:application/forms application))]
-      (assert form
-              (str "Form " form-id " not found in application " (:application/id application)))
-      (->> form
-           :form/fields
-           (mapv :field/title)
-           (mapv text/localized))))
+    (assert form
+            (str "Form " form-id " not found in application " (:application/id application)))
+    (->> form
+         :form/fields
+         (mapv :field/title)
+         (mapv text/localized))))
 
-(defn applications-to-csv [applications form-id user-id & {:keys [include-drafts]}]
-  (let [language (:language (user-settings/get-user-settings user-id))
-        applications (filter #(or include-drafts
-                                  (not= (:application/state %) :application.state/draft))
-                             applications)]
+(defn applications-to-csv [applications form-id user-id]
+  (let [language (:language (user-settings/get-user-settings user-id))]
     (if (empty? applications)
       ""
       (text/with-language language
         #(print-to-csv :column-names (concat (mapv (comp text/text :name) application-columns)
                                              (form-field-names form-id (first applications)))
                        :rows (mapv (partial application-to-row form-id) applications)
-                       :quote-strings? true)))))
+                       :quote-strings? true
+                       :strip-line-returns? true)))))
 
 (defn applications-filename []
   (format "applications_%s.csv" (str/replace (text/localize-time (time/now)) " " "_")))