+++ src/cljs/rems/actions/approve_reject.cljs
@@ -1,25 +1,37 @@
 (ns rems.actions.approve-reject
-  (:require [re-frame.core :as rf]
+  (:require [cljs-time.core :as time]
+            [cljs-time.format :as time-format]
+            [re-frame.core :as rf]
             [rems.actions.action :refer [action-attachment action-button action-comment action-form-view button-wrapper command!]]
+            [rems.atoms :refer [close-symbol]]
             [rems.text :refer [text]]))
 
 (def ^:private action-form-id "approve-reject")
 
+(defn default-end [length]
+  (when length
+    (time-format/unparse (time-format/formatters :year-month-day)
+                         (time/plus (time/now) (time/days length)))))
+
 (rf/reg-event-fx
  ::open-form
  (fn [{:keys [db]} _]
-   {:db (assoc db ::comment "")
+   {:db (assoc db ::comment "" ::entitlement-end (default-end (get-in db [:config :entitlement-default-length-days])))
     :dispatch [:rems.actions.action/set-attachments action-form-id []]}))
 
 (rf/reg-sub ::comment (fn [db _] (::comment db)))
 (rf/reg-event-db ::set-comment (fn [db [_ value]] (assoc db ::comment value)))
 
+(rf/reg-sub ::entitlement-end (fn [db _] (::entitlement-end db)))
+(rf/reg-event-db ::set-entitlement-end (fn [db [_ value]] (assoc db ::entitlement-end value)))
+
 (rf/reg-event-fx
  ::send-approve
- (fn [_ [_ {:keys [application-id comment attachments on-finished]}]]
+ (fn [_ [_ {:keys [application-id comment attachments end on-finished]}]]
    (command! :application.command/approve
              {:application-id application-id
               :comment comment
+              :entitlement-end (js/Date. end)
               :attachments attachments}
              {:description [text :t.actions/approve]
               :collapse action-form-id
@@ -45,7 +57,7 @@
                   :on-click #(rf/dispatch [::open-form])}])
 
 (defn approve-reject-view
-  [{:keys [application-id comment on-set-comment on-approve on-reject]}]
+  [{:keys [application-id comment on-set-comment end on-set-entitlement-end on-approve on-reject]}]
   [action-form-view action-form-id
    (text :t.actions/approve-reject)
    [[button-wrapper {:id "reject"
@@ -62,17 +74,36 @@
                      :comment comment
                      :on-comment on-set-comment}]
     [action-attachment {:application-id application-id
-                        :key action-form-id}]]])
+                        :key action-form-id}]
+    [:div.form-group
+     [:label {:for "approve-end"} (text :t.actions/approve-end-date)]
+     [:div.input-group.w-50
+      [:input.form-control {:type "date"
+                            :id "approve-end"
+                            :name "approve-end"
+                            :value end
+                            :required false
+                            :on-change #(on-set-entitlement-end (.. % -target -value))}]
+      (when end
+        [:div.input-group-append
+         [:button.btn.btn-outline-secondary
+          {:on-click #(on-set-entitlement-end nil)
+           :aria-label (text :t.actions/clear)}
+          [close-symbol]]])]]]])
 
 (defn approve-reject-form [application-id on-finished]
   (let [comment @(rf/subscribe [::comment])
-        attachments @(rf/subscribe [:rems.actions.action/attachments action-form-id])]
+        attachments @(rf/subscribe [:rems.actions.action/attachments action-form-id])
+        end @(rf/subscribe [::entitlement-end])]
     [approve-reject-view {:application-id application-id
                           :comment comment
                           :on-set-comment #(rf/dispatch [::set-comment %])
+                          :end end
+                          :on-set-entitlement-end #(rf/dispatch [::set-entitlement-end %])
                           :on-approve #(rf/dispatch [::send-approve {:application-id application-id
                                                                      :comment comment
                                                                      :attachments attachments
+                                                                     :end end
                                                                      :on-finished on-finished}])
                           :on-reject #(rf/dispatch [::send-reject {:application-id application-id
                                                                    :comment comment