+++ src/cljs/rems/administration/organization.cljs
@@ -36,7 +36,8 @@
 (rf/reg-sub ::loading? (fn [db _] (::loading? db)))
 
 (defn edit-button [id]
-  [atoms/link {:class "btn btn-primary edit-organization"}
+  [atoms/link {:id "edit-organization"
+               :class "btn btn-primary"}
    (str "/administration/organizations/edit/" id)
    (text :t.administration/edit)])
 
@@ -70,13 +71,12 @@
                                                                       (map enrich-user)
                                                                       (map :display)
                                                                       (interpose [:br]))]
-              [info-field
-               (text :t.administration/review-emails)
-               (->> (:organization/review-emails organization)
-                    (map display-localized-review-email)
-                    (interpose [:br]))
-               {:box? false
-                :solid? true}]
+              [:div
+               [:label (text :t.administration/review-emails)]
+               [:div.solid-group
+                (->> (:organization/review-emails organization)
+                     (map display-localized-review-email)
+                     (interpose [:br]))]]
               [inline-info-field (text :t.administration/active) [readonly-checkbox {:value (status-flags/active? organization)}]]
               [inline-info-field (text :t.administration/last-modified) (localize-time (:organization/last-modified organization))]
               [inline-info-field (text :t.administration/modifier) (:userid (:organization/modifier organization))]]}]